Added "Fixup whitespace on save" action
This runs a sed script to remove trailing whitespace from non-empty lines,
and to ensure that a newline exists at the end of the file. It is triggered
whenever a .swift file is saved.

Assorted minor cleanup: added `XCTAssertTrue` to the XCTAssert completions, cleaned trailing whitespace from "Test case class" snippet.